requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
crewai>=0.11.0
crewai-tools>=0.0.11
chromadb>=0.4.22
pydantic>=2.5.2
python-dotenv>=1.0.0
openai>=1.3.7
numpy>=1.26.2
pandas>=2.1.3 